% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iframe.R
\name{iframe}
\alias{iframe}
\title{Add an iframe
An iframe can contain a video for additional instructions within a Modal object}
\usage{
iframe(width, height, url_link)
}
\arguments{
\item{width}{a object of class character}

\item{height}{a object of class character}

\item{url_link}{a object of class character, indicating content to display}
}
\value{
a HTML object to be included in the ui section of a shiny app
}
\description{
Add an iframe
An iframe can contain a video for additional instructions within a Modal object
}
\examples{
iframe(width = "560", height = "315",
url_link = "https://www.youtube.com/embed/0fKg7e37bQE")
}
\seealso{
\href{http://getbootstrap.com/components/#responsive-embed}{Responsive Embed}
}
