% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_group.R
\name{list_group}
\alias{list_group}
\title{Add a Single List Group Component
A flexible method for displaying simple list elements as a group.}
\usage{
list_group(list_items)
}
\arguments{
\item{list_items}{a vector of content generated from the list_item function}
}
\value{
a HTML object to be included in the ui section of a shiny app
}
\description{
Add a Single List Group Component
A flexible method for displaying simple list elements as a group.
}
\examples{
list_group(list_item("Howdy Title", badge_value = 27))
}
\seealso{
\href{http://getbootstrap.com/components/#list-group}{List Group}
}
