% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_item.R
\name{list_item}
\alias{list_item}
\title{Add a Single List Item Component
A flexible method for displaying simple list elements.}
\usage{
list_item(content, badge = TRUE, badge_value)
}
\arguments{
\item{content}{a object of class character}

\item{badge}{a object of class logical}

\item{badge_value}{if the badge object is true provide an arguement object of
class numeric, named badge_value.}
}
\value{
a HTML object to be included in the ui section of a shiny app
}
\description{
Add a Single List Item Component
A flexible method for displaying simple list elements.
}
\examples{
list_item("Howdy Title", badge_value = 27)
}
\seealso{
\href{http://getbootstrap.com/components/#list-group}{List Group}
}
