% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_div.R
\name{panel_div}
\alias{panel_div}
\title{Add Panel Sections with Color Headers}
\usage{
panel_div(class_type, panel_title, content)
}
\arguments{
\item{class_type}{a object of class character, indicating the contextual state class}

\item{panel_title}{a object of class character, indicating the header title}

\item{content}{a object of class character, indicating text content}
}
\value{
a HTML object to be included in the ui section of a shiny app
}
\description{
Add Panel Sections with Color Headers
}
\seealso{
\href{http://getbootstrap.com/components/#panels-heading}{Bootstrap Panels}
}

