% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelDiv.R
\name{panelDiv}
\alias{panelDiv}
\title{Add Panel Sections with Color Headers}
\usage{
panelDiv(class_type, panel_title, content)
}
\arguments{
\item{class_type}{a object of class character, indicating the contextual state class}

\item{panel_title}{a object of class character, indicating the header title}

\item{content}{a object of class character, indicating text content}
}
\value{
HTML code to be applied to the ui section of a Shiny app
}
\description{
Add Panel Sections with Color Headers
}
\seealso{
\href{http://getbootstrap.com/components/#panels-heading}{Bootstrap Panels}
}

