% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run The Example Shiny App}
\usage{
runExample()
}
\value{
a shiny app lauched locally displaying example layouts using the
    shinyLP package functions and examples with the shinyBS package.
}
\description{
Run The Example Shiny App
}

