% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLP.R
\docType{package}
\name{shinyLP}
\alias{shinyLP-package}
\alias{shinyLP}
\title{A R package To Add Bootstrap Components to Shiny Applications
for Landing Home Pages.}
\description{
Provides additional Bootstrap web
framework components to enahance the user expereince and interface
for shiny applications. shinyLP makes it easier for R developers to include
advanced HTML components without the need for writing HTML. Landing home
pages are meant to be a informative introduction to the application,
its features and how the user can best use the tool.
}
\section{This package has functions for}{

  `iframe()`, `list_group()`, `list_item()`, `wells()`
  `jumbotron()`, `panel_div()`, and `thumbnail_label()`.
  Run the example shiny app with shinyLP::runExample().
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jasdumas/shinyLP}
}

}
\author{
\strong{Maintainer}: Jasmine Daly \email{jasmine.dumas@gmail.com}

}
