% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumbnail_label.R
\name{thumbnail_label}
\alias{thumbnail_label}
\title{Add a custom thumbnail with a label and button}
\usage{
thumbnail_label(image, label, content, button_link, button_label)
}
\arguments{
\item{image}{a object of class character}

\item{label}{a object of class character}

\item{content}{a object of class character}

\item{button_link}{a object of class character}

\item{button_label}{a object of class character}
}
\value{
a HTML object to be included in the ui section of a shiny app
}
\description{
Add a custom thumbnail with a label and button
}
\examples{
 thumbnail_label(image = 'Rlogo.png', label = 'App 1', content = 'here we go again',
 button_link = 'http://getbootstrap.com/', button_label = 'Click me')
}
\seealso{
\href{https://getbootstrap.com/components/#thumbnails-custom-content}{Thumbnails}
}
