% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wells.R
\name{wells}
\alias{wells}
\title{Add a Well Component
Use the well as a simple effect on an element to give it an inset effect.}
\usage{
wells(content, size = "default")
}
\arguments{
\item{content}{a object of class character}

\item{size}{a object of class character determining the size of the well
and options include, default, small, and large.}
}
\value{
a HTML object to be included in the ui section of a shiny app
}
\description{
Add a Well Component
Use the well as a simple effect on an element to give it an inset effect.
}
\examples{
wells(content = "Look, I'm a well!", size = "large")
}
\seealso{
\href{https://getbootstrap.com/components/#wells}{Wells}
}
